floats everywhere!

+++ src/mat_mult.cu
@@ -4,10 +4,10 @@
 // C(m,n) = A(m,k) * B(k,n)
 void gpu_blas_mmult(const double *A, const double *B, double *C, const int m, const int k, const int n) {
   int lda=m,ldb=k,ldc=m;
-  const float alf = 1;
-  const float bet = 0;
-  const float *alpha = &alf;
-  const float *beta = &bet;
+  const double alf = 1;
+  const double bet = 0;
+  const double *alpha = &alf;
+  const double *beta = &bet;
   // Create a handle for CUBLAS
   cublasHandle_t handle;
   cublasCreate(&handle);
